Updated Plugin: hyper-cache

+++ plugins/hyper-cache/plugin.php
@@ -4,12 +4,16 @@
   Plugin Name: Hyper Cache
   Plugin URI: http://www.satollo.net/plugins/hyper-cache
   Description: A easy to configure and efficient cache to increase the speed of your blog.
-  Version: 3.1.2
+  Version: 3.1.6
   Author: Stefano Lissa
   Author URI: http://www.satollo.net
   Disclaimer: Use at your own risk. No warranty expressed or implied is provided.
   Contributors: satollo
  */
+
+if (!defined('HYPER_CACHE_LOG')) {
+    define('HYPER_CACHE_LOG', false);
+}
 
 if (isset($_GET['cache'])) {
     if ($_GET['cache'] === '0') {
@@ -35,6 +39,7 @@ class HyperCache {
 
     var $post_id;
     var $options;
+    var $ob_started = false;
     static $instance;
 
     const MOBILE_AGENTS = 'android|iphone|iemobile|up.browser|up.link|mmp|symbian|smartphone|midp|wap|phone|ipod|xoom|blackberry';
@@ -47,6 +52,7 @@ class HyperCache {
         register_deactivation_hook('hyper-cache/plugin.php', array($this, 'hook_deactivate'));
 
         add_action('edit_post', array($this, 'hook_edit_post'), 1);
+        add_action('save_post', array($this, 'hook_save_post'), 1);
         add_action('comment_post', array($this, 'hook_comment_post'), 1, 2);
         add_action('wp_update_comment_count', array($this, 'hook_wp_update_comment_count'), 1);
         add_action('bbp_new_reply', array($this, 'hook_bbp_new_reply'));
@@ -69,10 +75,25 @@ class HyperCache {
             add_action('admin_init', array($this, 'hook_admin_init'));
             add_action('admin_head', array($this, 'hook_admin_head'));
             add_action('admin_menu', array($this, 'hook_admin_menu'));
+
+            if (!is_dir(WP_CONTENT_DIR . '/logs')) {
+                wp_mkdir_p(WP_CONTENT_DIR . '/logs');
+            }
         }
     }
 
+    function log($object) {
+        if (is_object($object) || is_array($object)) {
+            $object = print_r($object, true);
+        }
+        //file_put_contents(WP_CONTENT_DIR . '/logs/hyper-cache.log', $object, FILE_APPEND);
+        error_log($object);
+    }
+
     function hook_activate() {
+        if (!is_dir(WP_CONTENT_DIR . '/logs')) {
+            wp_mkdir_p(WP_CONTENT_DIR . '/logs');
+        }
 
         if (!isset($this->options['mobile'])) {
             $this->options['mobile'] = 0;
@@ -166,6 +187,10 @@ class HyperCache {
         $advanced_cache = str_replace('HC_BROWSER_CACHE', isset($this->options['browser_cache']) ? 1 : 0, $advanced_cache);
 
         $advanced_cache = str_replace('HC_HTTPS', (int) $this->options['https'], $advanced_cache);
+        $advanced_cache = str_replace('HC_READFILE', isset($this->options['readfile']) ? 1 : 0, $advanced_cache);
+
+        $advanced_cache = str_replace('HC_SERVE_EXPIRED_TO_BOT', isset($this->options['serve_expired_to_bots']) ? 1 : 0, $advanced_cache);
+        $advanced_cache = str_replace('HC_BOTS_IGNORE_NOCACHE', isset($this->options['bots_ignore_nocache']) ? 1 : 0, $advanced_cache);
 
         return file_put_contents(WP_CONTENT_DIR . '/advanced-cache.php', $advanced_cache);
     }
@@ -198,32 +223,68 @@ class HyperCache {
     }
 
     function hook_comment_post($comment_id, $status) {
+        if (HYPER_CACHE_LOG)
+            $this->log('Hook: comment_post');
         if ($status === 1) {
             $comment = get_comment($comment_id);
-            $this->hook_edit_post($comment->comment_post_ID);
+            $this->clean_post($comment->comment_post_ID, isset($this->options['clean_archives_on_comment']), isset($this->options['clean_archives_on_comment']));
         }
     }
 
     function hook_wp_update_comment_count($post_id) {
+        if (HYPER_CACHE_LOG)
+            $this->log('Hook: wp_update_comment_count');
         if ($this->post_id == $post_id) {
             return;
         }
-        $this->hook_edit_post($post_id);
+        $this->clean_post($post_id, isset($this->options['clean_archives_on_comment']), isset($this->options['clean_home_on_comment']));
     }
 
+
+    function hook_save_post($post_id) {
+        if (HYPER_CACHE_LOG)
+            $this->log('Hook: save_post');
+    }
+
+    /**
+     * edit_post is called even when a comment is added, but the comment hook prevent the execution of
+     * edit_post like if the post has been modified.
+     */
     function hook_edit_post($post_id) {
+        if (HYPER_CACHE_LOG)
+            $this->log('Hook: edit_post');
+        $this->clean_post($post_id, isset($this->options['clean_archives_on_post_edit']), isset($this->options['clean_home_on_post_edit']));
+    }
+
+    function clean_post($post_id, $clean_archives = true, $clean_home = true) {
+        if (HYPER_CACHE_LOG)
+            $this->log('Cleaning post ' . $post_id);
         // When someone deletes the advaced-cache.php file
         if (!function_exists('hyper_cache_sanitize_uri')) {
+            if (HYPER_CACHE_LOG)
+                $this->log('hyper_cache_sanitize_uri does not exists');
             return;
         }
 
         if ($this->post_id == $post_id) {
+            if (HYPER_CACHE_LOG)
+                $this->log('Already cleaned in this session');
             return;
         }
 
-        if (get_post_status($post_id) != 'publish') {
-            //error_log('Not a published post');
+        $status = get_post_status($post_id);
+        if (HYPER_CACHE_LOG)
+                $this->log('Status: ' . $status);
+        if ($status != 'publish' && $status != 'trash') {
+            if (HYPER_CACHE_LOG)
+                $this->log('Post not published');
+            //$this->log('Not a published post');
             return;
+        }
+
+        if ($status == 'trash') {
+            $clean_archives = true;
+            $clean_home = true;
         }
 
         $this->post_id = $post_id;
@@ -239,55 +300,68 @@ class HyperCache {
             }
         }
 
-        $dir = $folder . '/' . substr(get_option('home'), strpos(get_option('home'), '://') + 3);
 
-        // The home
-        @unlink($dir . '/index.html');
-        @unlink($dir . '/index.html.gz');
-        @unlink($dir . '/index-https.html');
-        @unlink($dir . '/index-https.html.gz');
-        @unlink($dir . '/index-mobile.html');
-        @unlink($dir . '/index-mobile.html.gz');
-        @unlink($dir . '/index-https-mobile.html');
-        @unlink($dir . '/index-https-mobile.html.gz');
+        $dir = $folder . substr(get_option('home'), strpos(get_option('home'), '://') + 3);
 
-        $this->remove_dir($dir . '/feed/');
-        // Home subpages
-        $this->remove_dir($dir . '/page/');
+        if ($clean_home) {
+            if (HYPER_CACHE_LOG)
+                $this->log('Cleaning the home');
 
-        @unlink($dir . '/robots.txt');
+            // The home
+            if (HYPER_CACHE_LOG)
+                $this->log('Cleaning the home index*');
+            @unlink($dir . '/index.html');
+            @unlink($dir . '/index.html.gz');
+            @unlink($dir . '/index-https.html');
+            @unlink($dir . '/index-https.html.gz');
+            @unlink($dir . '/index-mobile.html');
+            @unlink($dir . '/index-mobile.html.gz');
+            @unlink($dir . '/index-https-mobile.html');
+            @unlink($dir . '/index-https-mobile.html.gz');
 
-        $base = get_option('category_base');
-        if (empty($base)) {
-            $base = 'category';
+            $this->remove_dir($dir . '/feed/');
+            // Home subpages
+            $this->remove_dir($dir . '/page/');
         }
-        $this->remove_dir($dir . '/' . $base . '/');
 
-        $permalink_structure = get_option('permalink_structure');
-        //error_log(substr($permalink_structure, 0, 11));
-        if (substr($permalink_structure, 0, 11) == '/%category%') {
-            $categories = get_categories();
-            //error_log(print_r($categories, true));
-            foreach ($categories as &$category) {
-                //error_log('Removing: ' . $dir . '/' . $category->slug . '/');
-                $this->remove_page($dir . '/' . $category->slug);
+        //@unlink($dir . '/robots.txt');
+        if ($clean_archives) {
+            if (HYPER_CACHE_LOG)
+                $this->log('Cleaning archives');
+
+            $base = get_option('category_base');
+            if (empty($base)) {
+                $base = 'category';
             }
+            $this->remove_dir($dir . '/' . $base . '/');
+
+            $permalink_structure = get_option('permalink_structure');
+            //$this->log(substr($permalink_structure, 0, 11));
+            if (substr($permalink_structure, 0, 11) == '/%category%') {
+                $categories = get_categories();
+                //$this->log(print_r($categories, true));
+                foreach ($categories as &$category) {
+                    //$this->log('Removing: ' . $dir . '/' . $category->slug . '/');
+                    $this->remove_page($dir . '/' . $category->slug);
+                }
+            }
+
+            $base = get_option('tag_base');
+            if (empty($base)) {
+                $base = 'tag';
+            }
+            $this->remove_dir($dir . '/' . $base . '/');
+
+            $this->remove_dir($dir . '/type/');
+
+            $this->remove_dir($dir . '/' . date('Y') . '/');
         }
-
-        $base = get_option('tag_base');
-        if (empty($base)) {
-            $base = 'tag';
-        }
-        $this->remove_dir($dir . '/' . $base . '/');
-
-        $this->remove_dir($dir . '/type/');
-
-        $this->remove_dir($dir . '/' . date('Y') . '/');
     }
 
     /*
      * Runs only if $hyper_cache_is_mobile is true
      */
+
     function hook_get_stylesheet($stylesheet = '') {
         $theme = wp_get_theme($this->options['theme']);
         if (!$theme->exists()) {
@@ -301,6 +375,7 @@ class HyperCache {
      *
      * var WP_Theme $theme
      */
+
     function hook_get_template($template) {
         $theme = wp_get_theme($this->options['theme']);
         if (!$theme->exists()) {
@@ -310,9 +385,17 @@ class HyperCache {
     }
 
     function hook_template_redirect() {
-        global $cache_stop, $hyper_cache_stop, $hyper_cache_stop;
+        global $cache_stop, $hyper_cache_stop, $lite_cache_stop;
 
-        if ($cache_stop || $hyper_cache_stop || $hyper_cache_stop) {
+        if (HYPER_CACHE_LOG) {
+            $this->log('hook_template_redirect');
+        }
+
+        if ($this->ob_started) {
+            return;
+        }
+
+        if ($cache_stop || $hyper_cache_stop || $lite_cache_stop) {
             return;
         }
 
@@ -341,7 +424,7 @@ class HyperCache {
             if (file_exists($file) && ($this->options['max_age'] == 0 || filemtime($file) > time() - $this->options['max_age'] * 3600)) {
                 header('Content-Type: text/html;charset=UTF-8');
                 // For some reason it seems more performant than readfile...
-                header('X-Hyper-Cache: hit');
+                header('X-Hyper-Cache: hit,404');
                 echo file_get_contents($file);
                 die();
             }
@@ -416,6 +499,7 @@ class HyperCache {
             }
         }
 
+        $this->ob_started = true;
         ob_start('hyper_cache_callback');
     }
 
@@ -426,6 +510,8 @@ class HyperCache {
     }
 
     function remove_page($dir) {
+        if (HYPER_CACHE_LOG)
+            $this->log('Removing page: ' . $dir);
         $dir = untrailingslashit($dir);
         @unlink($dir . '/index.html');
         @unlink($dir . '/index.html.gz');
@@ -442,7 +528,9 @@ class HyperCache {
     }
 
     function remove_dir($dir) {
-        //error_log('Removing dir: ' . $dir);
+        if (HYPER_CACHE_LOG) {
+            $this->log('Removing dir: ' . $dir);
+        }
         $dir = trailingslashit($dir);
         $files = glob($dir . '*', GLOB_MARK);
         if (!empty($files)) {
@@ -458,10 +546,19 @@ class HyperCache {
     }
 
     function hook_hyper_cache_clean() {
-        //error_log('hook_hyper_cache_clean');
-        if ($this->options['max_age'] == 0)
+        if (!isset($this->options['autoclean'])) {
             return;
+        }
+        if (HYPER_CACHE_LOG) {
+            $this->log('Periodic cache cleaning start');
+        }
+        if ($this->options['max_age'] == 0) {
+            return;
+        }
         $this->remove_older_than(time() - $this->options['max_age'] * 3600);
+        if (HYPER_CACHE_LOG) {
+            $this->log('Periodic cache cleaning stop');
+        }
     }
 
     function remove_older_than($time) {
@@ -475,9 +572,9 @@ class HyperCache {
                 if (substr($file, -1) == '/')
                     $this->_remove_older_than($time, $file);
                 else {
-                    //error_log($file . ' ' . ($time-filemtime($file)));
+                    //$this->log($file . ' ' . ($time-filemtime($file)));
                     if (@filemtime($file) < $time) {
-                        //error_log('Removing ' . $file);
+                        //$this->log('Removing ' . $file);
                         @unlink($file);
                     }
                 }
@@ -525,10 +622,13 @@ class HyperCache {
 function hyper_cache_callback($buffer) {
     global $cache_stop, $lite_cache, $hyper_cache_stop, $hyper_cache_group;
 
-    if ($cache_stop || $hyper_cache_stop)
+    if ($cache_stop || $hyper_cache_stop) {
         return $buffer;
-    if (strlen($buffer) == 0)
+    }
+
+    if (strlen($buffer) == 0) {
         return '';
+    }
 
     $uri = hyper_cache_sanitize_uri($_SERVER['REQUEST_URI']);
 
@@ -571,7 +671,7 @@ function hyper_cache_callback($buffer) {
         $script .= 'var z = lc_get_cookie("comment_author_email_' . COOKIEHASH . '");';
         $script .= 'if (z != "") e["email"].value = z;';
         $script .= 'z = lc_get_cookie("comment_author_' . COOKIEHASH . '");';
-        $script .= 'if (z != "") e["author"].value = z;';
+        $script .= 'if (z != "") e["author"].value = z.replace(/\+/g, " ");';
         $script .= 'z = lc_get_cookie("comment_author_url_' . COOKIEHASH . '");';
         $script .= 'if (z != "") e["url"].value = z;';
         $script .= '}';
@@ -601,13 +701,22 @@ function hyper_cache_callback($buffer) {
 if (!function_exists('hyper_cache_sanitize_uri')) {
 
     function hyper_cache_sanitize_uri($uri) {
-        $uri = preg_replace('/[^a-zA-Z0-9\.\/\-_]+/', '_', $uri);
+        if (HYPER_CACHE_LOG) {
+            $this->log('URI: ' . $uri);
+        }
+        $uri = preg_replace('/[^a-zA-Z0-9\/\-_]+/', '_', $uri);
         $uri = preg_replace('/\/+/', '/', $uri);
-        $uri = rtrim($uri, '.-_/');
+        //$uri = rtrim($uri, '.-_/');
         if (empty($uri) || $uri[0] != '/') {
             $uri = '/' . $uri;
         }
-        return rtrim($uri, '/');
+        if (strlen($uri) > 1 && substr($uri, -1, 1) == '/') {
+            $uri = rtrim($uri, '/') . '_';
+        }
+        if (HYPER_CACHE_LOG) {
+            $this->log('Sanitized URI: ' . $uri);
+        }
+        return $uri;
     }
 
 }
